+++ deep4/prototype/lobaio.py
@@ -23,28 +23,28 @@ def get_trees(request, special_mode=None):
     return render(request, f"prototype/{template_sub_dir}root.html", context)
 
 
-def get_tree(request, location, special_mode=None):
-    model_class = Location
-    loc = model_class.objects.get(id=location)
-    context = {"location": loc, "content_object": loc.content_object}
-    template_sub_dir = ""
-    if special_mode == "help":
-        context.update({"help": context["location"].content_object.__class__.__doc__})
-    elif special_mode == "debug":
-        context.update({"db": f"loc={context['location'].id}; content= {context['location'].content_object.id}"})
-    elif special_mode == "hide":
-        print("hide button")
-
-    if request.htmx and not request.POST:
-        template_name = "prototype/tree.html"
-    elif request.POST:
-        template_name = f"prototype/{location.content_type.model}content_display.html"
-    else:
-        context.update({"location_list": [context["location"]]})
-        template_name = f"prototype/{template_sub_dir}tree.html"
-
-    print(f"{context=}-{template_name=}")
-    return render(request, template_name, context)
+# def get_tree(request, location, special_mode=None):
+#     model_class = Location
+#     loc = model_class.objects.get(id=location)
+#     context = {"location": loc, "content_object": loc.content_object}
+#     template_sub_dir = ""
+#     if special_mode == "help":
+#         context.update({"help": context["location"].content_object.__class__.__doc__})
+#     elif special_mode == "debug":
+#         context.update({"db": f"loc={context['location'].id}; content= {context['location'].content_object.id}"})
+#     elif special_mode == "hide":
+#         print("hide button")
+#
+#     if request.htmx and not request.POST:
+#         template_name = "prototype/tree_ol.html"
+#     elif request.POST:
+#         template_name = f"prototype/{location.content_type.model}content_display.html"
+#     else:
+#         context.update({"location_list": [context["location"]]})
+#         template_name = f"prototype/{template_sub_dir}tree_ol.html"
+#
+#     print(f"{context=}-{template_name=}")
+#     return render(request, template_name, context)
 
 
 class PlainTextForm(forms.ModelForm):
@@ -66,19 +66,16 @@ class RichTextForm(forms.ModelForm):
         fields = [
             "content",
         ]
-        widgets = {"content": Textarea(attrs={"hidden": True})}
+        widgets = {"content": Textarea(attrs={"cols": 70, "rows": 1, "hidden": True})}
 
 
 FORMS = {"location": None, "plaintext": PlainTextForm, "richtext": RichTextForm}
 
 
-class ContentModelFormView(View):
-    """The FormView for the SimpleTextForm:
-    provides actions for
-    get - returns an editable form
-    post - updates the value
-    button actions for
-          add child, add sibling and delete nodes."""
+class LocationView(View):
+    """
+    The LocationViews, these are anchored only at the location.
+    """
 
     def setup(self, request, *args, **kwargs):
         super().setup(request, *args, **kwargs)
@@ -89,39 +86,42 @@ class ContentModelFormView(View):
             self.model_class = self.content.__class__
             self.model_form_class = FORMS[self.location.content_type.model]
             self.template = self.content.default_template
-            # self.context = add_to_context(self.context, [self.location, self.content])
-        elif kwargs.get("content_type"):
-            self.model_class = ContentType.objects.get(
-                app_label="prototype", model=kwargs.get("content_type")
-            ).model_class()
-            if kwargs.get("content"):
-                self.content = self.model_class.objects.get(id=kwargs.get("content"))
-            self.template = self.model_class.default_template
-            self.model_form_class = FORMS[kwargs.get("content_type")]
-            self.location_model_class = Location
-        else:
-            self.location_model_class = Location
 
-    def get(self, request, content_type, content, special_mode=None):
-        """gets a form that holds the editable content"""
-        match special_mode:
-            case "help":
-                return get_tree(request, self.location.id, "help")
-            case "debug":
-                return get_tree(request, self.location.id, "db")
-            case _:
-                stf = self.model_form_class(instance=self.content)
-        return render(request, self.model_form_class.template_name, {"form": stf})
+    def get(self, request, location=None, special_mode=None):
+        # something to do get_tree and get_trees
+        if location:
+            loc = Location.objects.get(id=location)
+            context = {"location": loc, "content_object": loc.content_object}
+        else:
+            loc = Location.get_root_nodes()
+            context = {"location": loc, "content_object": None}
+        template_sub_dir = ""
+        if special_mode == "help":
+            context.update({"help": context["location"].content_object.__class__.__doc__})
+        elif special_mode == "debug":
+            context.update({"db": f"loc={context['location'].id}; content= {context['location'].content_object.id}"})
+        elif special_mode == "hide":
+            print("hide button")
+
+        if request.htmx and not request.POST:
+            template_name = "prototype/tree_ol.html"
+        elif request.POST:
+            template_name = f"prototype/{location.content_type.model}content_display.html"
+        else:
+            context.update({"location_list": [context["location"]]})
+            template_name = f"prototype/{template_sub_dir}root.html"
+        print(f"{context=}-{template_name=}")
+        return render(request, template_name, context)
 
     def patch(self, request, location, special_mode=None):
         """ "add child location to location, provide default content."""
-        pt_instance = self.model_class.objects.create(content="new entry")
+        pt_instance = self.model_class.objects.create(content="new child")
         self.location.add_child(content_object=pt_instance)
-        return get_tree(request, self.location.id)
+        return self.get(request, self.location.id)
 
-    def put(self, request, location=None, content_type="plaintext", special_mode=None):
+    def put(self, request, location=None, special_mode=None):
         """add a sibling node at this depth"""
-        content_instance = self.model_class.objects.create(content="new entry")
+        content_instance = self.model_class.objects.create(content="new sibling")
         if location:
             loc_instance = self.location_model_class.objects.get(id=location)
             if loc_instance.is_root():
@@ -131,7 +131,37 @@ class ContentModelFormView(View):
         else:
             new_loc = self.location_model_class.add_root(content_object=content_instance)
 
-        return get_tree(request, new_loc.id)
+        return self.get(request, new_loc.id)
+
+    def delete(self, request, location):
+        """delete this node and the nodes below it."""
+        # I believe this deletes sub-locations but maybe not content  ??
+        self.location.delete()
+        return HttpResponse("")
+
+
+class ContentModelFormView(View):
+    """The FormViews for simple & rich content forms:
+    provides actions for
+    get - returns an editable form
+    post - updates the value and returns the tree
+    """
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        if kwargs.get("content_type"):
+            self.model_class = ContentType.objects.get(
+                app_label="prototype", model=kwargs.get("content_type")
+            ).model_class()
+            if kwargs.get("content"):
+                self.content = self.model_class.objects.get(id=kwargs.get("content"))
+            self.template = self.model_class.default_template
+            self.model_form_class = FORMS[kwargs.get("content_type")]
+
+    def get(self, request, content_type, content, special_mode=None):
+        """gets a form that holds the editable content"""
+        stf = self.model_form_class(instance=self.content)
+        return render(request, self.model_form_class.template_name, {"form": stf})
 
     def post(self, request, content_type, content, special_mode=None):
         """update the content of the node."""
@@ -147,14 +177,20 @@ class ContentModelFormView(View):
         else:
             print(content_form.errors)
 
-    def delete(self, request, location):
-        """delete this node and the nodes below it."""
-        # I believe this deletes sub-locations but maybe not content  ??
-        self.location.delete()
-        return HttpResponse("")
+
+def page_view(request, location, page_no):
+    project = Location.objects.get(id=location)
+    if not project.is_root():
+        return "Location is not a project."
+    pages = project.get_children()
+    if page_no > len(pages):
+        return "Invalid page number."
+    return LocationView.as_view()(request, location=pages[page_no].id)
 
 
-def add_to_context(context, *variables):
-    for v in variables:
-        context.update({f"{v=}".split("=")[0]: v})
-    return context
+def project_view(request, location):
+    project = Location.objects.get(id=location)
+    if not project.is_root():
+        return "Location is not a project."
+    project.get_children()  # refacto
+    return LocationView.as_view()(request, location)