+++ deep4/prototype/static/jqui-htmx.js
@@ -64,29 +64,46 @@ $(document).on('htmx:afterSwap', function(event) {
 function add_jq(event) {
   console.log("afterSwap", event);
   let elmt = $(event.target);
-  elmt.find('[data-jqui]').each(function () {
-    console.log("this 1", $(this)[0], $(this)[0].getAttribute("data-jqui"), $(this)[0]["dataset"]);
+  console.log("find jqui", elmt.find('[data-jqui]'));
+  console.log(elmt.find('[data-jqui]').each(function(e) {console.log("item", e)}));
+  elmt.find('[data-jqui]').each(function (item) {
+      let element =
+    console.log("this 1", $(this))
     let id_ = `#${$(this)[0].getAttribute("id")}`;
     let interactions = $(this)[0].getAttribute("data-jqui");
     let interaction_options = $(this)[0].getAttribute("data-jqui-opts");
+    console.log("adding", interactions, "with", interaction_options, "to", id_)
+
     if ((interactions) && (interaction_options)) {
       console.log("options", interaction_options.split(" "));
       $(id_)[interactions]()[interactions](...interaction_options.split(" "));
     }
-    else if (interactions) {
+    // if ((interactions) && (interaction_options)) {
+    //   console.log("options", interaction_options.split(" "));
+    //   $(id_)[interactions]()[interactions](...interaction_options.split(" "));
+    // }
+    if (interactions) {
+        console.log("adding", interactions, "to", id_)
       $(id_)[interactions]();
     };
   }  );
 }
 
-document.addEventListener('htmx:afterSwap', add_jq);
+// document.addEventListener('htmx:afterSwap', add_jq);
 document.addEventListener('DOMContentLoaded', add_jq);
 
 function add_position_updates(event) {
   console.log("size/position update", event.type);
   if (event.type === "mouseup"){
-     console.log("hit mouseup", event.target.parentNode.getAttribute("style"));
-     console.log("imagine we post:", event.target.parentNode.id, event.target.parentNode.getAttribute("style"), event.target);
+     //console.log("hit mouseup", event.target.parentNode.getAttribute("style"));
+     //console.log("imagine we post:", event.target.parentNode.id, event.target.parentNode.getAttribute("style"), event.target);
+    console.log("event", event)
+     let left_of_self = event.target.previousElementSibling;
+     let right_of_self = event.target.nextElementSibling;
+
+     console.log("sortable - post:\n self", event.target,
+         "\n left-", left_of_self,
+         "\n right-", right_of_self);
   }
 }
-// document.addEventListener('mouseup', add_position_updates, false)
+document.addEventListener('mouseup', add_position_updates, false)